Ekoslive multi-node support
Add support for connecting to Online/Offline servers at the same time. Can be extended to support any numbers of servers in the future.

+++ kstars/ekos/ekoslive/media.cpp
@@ -8,7 +8,6 @@
 
 #include "media.h"
 #include "commands.h"
-#include "profileinfo.h"
 #include "skymapcomposite.h"
 #include "fitsviewer/fitsview.h"
 #include "fitsviewer/fitsdata.h"
@@ -28,13 +27,19 @@
 namespace EkosLive
 {
 
-Media::Media(Ekos::Manager * manager): m_Manager(manager)
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
+Media::Media(Ekos::Manager * manager, QVector<QSharedPointer<NodeManager>> &nodeManagers):
+    m_Manager(manager), m_NodeManagers(nodeManagers)
 {
-    connect(&m_WebSocket, &QWebSocket::connected, this, &Media::onConnected);
-    connect(&m_WebSocket, &QWebSocket::disconnected, this, &Media::onDisconnected);
-    connect(&m_WebSocket, static_cast<void(QWebSocket::*)(QAbstractSocket::SocketError)>(&QWebSocket::error), this,
-            &Media::onError);
-
+    for (auto &nodeManager : m_NodeManagers)
+    {
+        connect(nodeManager->media(), &Node::connected, this, &Media::onConnected);
+        connect(nodeManager->media(), &Node::disconnected, this, &Media::onDisconnected);
+        connect(nodeManager->media(), &Node::onTextReceived, this, &Media::onTextReceived);
+        connect(nodeManager->media(), &Node::onBinaryReceived, this, &Media::onBinaryReceived);
+    }
 
     connect(this, &Media::newMetadata, this, &Media::uploadMetadata);
     connect(this, &Media::newImage, this, [this](const QByteArray & image)
@@ -44,78 +49,57 @@ Media::Media(Ekos::Manager * manager): m_Manager(manager)
     });
 }
 
-void Media::connectServer()
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
+bool Media::isConnected() const
 {
-    QUrl requestURL(m_URL);
-
-    QUrlQuery query;
-    query.addQueryItem("username", m_AuthResponse["username"].toString());
-    query.addQueryItem("token", m_AuthResponse["token"].toString());
-    if (m_AuthResponse.contains("remoteToken"))
-        query.addQueryItem("remoteToken", m_AuthResponse["remoteToken"].toString());
-    query.addQueryItem("cloudEnabled", Options::ekosLiveCloud() ? "true" : "false");
-    query.addQueryItem("email", m_AuthResponse["email"].toString());
-    query.addQueryItem("from_date", m_AuthResponse["from_date"].toString());
-    query.addQueryItem("to_date", m_AuthResponse["to_date"].toString());
-    query.addQueryItem("plan_id", m_AuthResponse["plan_id"].toString());
-    query.addQueryItem("type", m_AuthResponse["type"].toString());
-    query.addQueryItem("version", KSTARS_VERSION);
-
-    requestURL.setPath("/media/ekos");
-    requestURL.setQuery(query);
-
-
-    m_WebSocket.open(requestURL);
-
-    qCInfo(KSTARS_EKOS) << "Connecting to Websocket server at" << requestURL.toDisplayString();
+    return std::any_of(m_NodeManagers.begin(), m_NodeManagers.end(), [](auto & nodeManager)
+    {
+        return nodeManager->message()->isConnected();
+    });
 }
 
-void Media::disconnectServer()
-{
-    m_WebSocket.close();
-}
-
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::onConnected()
 {
-    qCInfo(KSTARS_EKOS) << "Connected to media Websocket server at" << m_URL.toDisplayString();
+    auto node = qobject_cast<Node*>(sender());
+    if (!node)
+        return;
 
-    connect(&m_WebSocket, &QWebSocket::textMessageReceived,  this, &Media::onTextReceived, Qt::UniqueConnection);
-    connect(&m_WebSocket, &QWebSocket::binaryMessageReceived, this, &Media::onBinaryReceived, Qt::UniqueConnection);
-
-    m_isConnected = true;
-    m_ReconnectTries = 0;
+    qCInfo(KSTARS_EKOS) << "Connected to Media Websocket server at" << node->url().toDisplayString();
 
     emit connected();
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::onDisconnected()
 {
-    qCInfo(KSTARS_EKOS) << "Disconnected from media Websocket server.";
-    m_isConnected = false;
+    auto node = qobject_cast<Node*>(sender());
+    if (!node)
+        return;
 
-    disconnect(&m_WebSocket, &QWebSocket::textMessageReceived,  this, &Media::onTextReceived);
-    disconnect(&m_WebSocket, &QWebSocket::binaryMessageReceived, this, &Media::onBinaryReceived);
+    qCInfo(KSTARS_EKOS) << "Disconnected from Message Websocket server at" << node->url().toDisplayString();
 
-    m_sendBlobs = true;
-
-    for (const QString &oneFile : temporaryFiles)
-        QFile::remove(oneFile);
-    temporaryFiles.clear();
-
-    emit disconnected();
-}
-
-void Media::onError(QAbstractSocket::SocketError error)
-{
-    qCritical(KSTARS_EKOS) << "Media Websocket connection error" << m_WebSocket.errorString();
-    if (error == QAbstractSocket::RemoteHostClosedError ||
-            error == QAbstractSocket::ConnectionRefusedError)
+    if (isConnected() == false)
     {
-        if (m_ReconnectTries++ < RECONNECT_MAX_TRIES)
-            QTimer::singleShot(RECONNECT_INTERVAL, this, SLOT(connectServer()));
+        m_sendBlobs = true;
+
+        for (const QString &oneFile : temporaryFiles)
+            QFile::remove(oneFile);
+        temporaryFiles.clear();
+
+        emit disconnected();
     }
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::onTextReceived(const QString &message)
 {
     qCInfo(KSTARS_EKOS) << "Media Text Websocket Message" << message;
@@ -206,6 +190,9 @@ void Media::onTextReceived(const QString &message)
     }
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::onBinaryReceived(const QByteArray &message)
 {
     // Sometimes this is triggered even though it's a text message
@@ -220,14 +207,20 @@ void Media::onBinaryReceived(const QByteArray &message)
     }
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::sendDarkLibraryData(const QSharedPointer<FITSData> &data)
 {
     sendData(data, "+D");
 };
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::sendData(const QSharedPointer<FITSData> &data, const QString &uuid)
 {
-    if (m_isConnected == false || Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
+    if (Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
         return;
 
     m_UUID = uuid;
@@ -237,9 +230,12 @@ void Media::sendData(const QSharedPointer<FITSData> &data, const QString &uuid)
     QtConcurrent::run(this, &Media::upload, m_TemporaryView);
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::sendFile(const QString &filename, const QString &uuid)
 {
-    if (m_isConnected == false || Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
+    if (Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
         return;
 
     m_UUID = uuid;
@@ -252,9 +248,12 @@ void Media::sendFile(const QString &filename, const QString &uuid)
     previewImage->loadFile(filename);
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::sendView(const QSharedPointer<FITSView> &view, const QString &uuid)
 {
-    if (m_isConnected == false || Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
+    if (Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
         return;
 
     m_UUID = uuid;
@@ -262,6 +261,9 @@ void Media::sendView(const QSharedPointer<FITSView> &view, const QString &uuid)
     upload(view);
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::upload(const QSharedPointer<FITSView> &view)
 {
     const QString ext = "jpg";
@@ -335,6 +337,9 @@ void Media::upload(const QSharedPointer<FITSView> &view)
     emit newImage(jpegData);
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::sendUpdatedFrame(const QSharedPointer<FITSView> &view)
 {
     QString ext = "jpg";
@@ -429,9 +434,12 @@ void Media::sendUpdatedFrame(const QSharedPointer<FITSView> &view)
     emit newImage(jpegData);
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::sendVideoFrame(const QSharedPointer<QImage> &frame)
 {
-    if (m_isConnected == false || Options::ekosLiveImageTransfer() == false || m_sendBlobs == false || !frame)
+    if (Options::ekosLiveImageTransfer() == false || m_sendBlobs == false || !frame)
         return;
 
     int32_t width = Options::ekosLiveHighBandwidth() ? HB_VIDEO_WIDTH : HB_VIDEO_WIDTH / 2;
@@ -462,14 +470,17 @@ void Media::sendVideoFrame(const QSharedPointer<QImage> &frame)
     writer.write(videoImage);
     buffer.close();
 
-    m_WebSocket.sendBinaryMessage(image);
+    for (auto &nodeManager : m_NodeManagers)
+    {
+        nodeManager->media()->sendBinaryMessage(image);
+    }
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////
+///
+///////////////////////////////////////////////////////////////////////////////////////////
 void Media::registerCameras()
 {
-    if (m_isConnected == false)
-        return;
-
     for(auto &oneDevice : INDIListener::devices())
     {
         auto camera = oneDevice->getCamera();
@@ -481,18 +492,23 @@ void Media::registerCameras()
 void Media::resetPolarView()
 {
     this->correctionVector = QLineF();
-
     m_Manager->alignModule()->zoomAlignView();
 }
 
 void Media::uploadMetadata(const QByteArray &metadata)
 {
-    m_WebSocket.sendTextMessage(metadata);
+    for (auto &nodeManager : m_NodeManagers)
+    {
+        nodeManager->media()->sendTextMessage(metadata);
+    }
 }
 
 void Media::uploadImage(const QByteArray &image)
 {
-    m_WebSocket.sendBinaryMessage(image);
+    for (auto &nodeManager : m_NodeManagers)
+    {
+        nodeManager->media()->sendBinaryMessage(image);
+    }
 }
 
 void Media::processNewBLOB(IBLOB *bp)
@@ -502,7 +518,7 @@ void Media::processNewBLOB(IBLOB *bp)
 
 void Media::sendModuleFrame(const QSharedPointer<FITSView> &view)
 {
-    if (m_isConnected == false || Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
+    if (Options::ekosLiveImageTransfer() == false || m_sendBlobs == false)
         return;
 
     if (qobject_cast<Ekos::Align*>(sender()) == m_Manager->alignModule())
